+++ Source/ACE/Factories/GenericObjectFactory.cs
@@ -51,12 +51,13 @@ namespace ACE.Factories
                         // Use the DebugObject to assist in building proper objects for weenies
                         // FIXME(ddevec): Some objects have null location.  This freaks out the landblock... ignore them?
                         if (aceO.Location != null)
+                        {
                             results.Add(new DebugObject(aceO));
+                        }
                         break;
 #endif
                 }
             }
-
             return results;
         }
     }